refactor: 패키지 구조 변경
- 계층형 구조에서 도메인형 구조로 변경
- 도메인은 purchase, draw, result 3개의 도메인 사용
