+++ src/test/java/lotto/draw/model/LottoTest.java
@@ -1,10 +1,10 @@
-package lotto.model.draw;
+package lotto.draw.model;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
 import java.util.List;
-import lotto.model.purchase.Lotto;
+import lotto.purchase.model.Lotto;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
